Add files via upload
wire colors with pin-out  from my E36325i M50NV 1992 for bosch M3.1 402 ECU (european car) 

This pinout was originaly made for fault finding on a Non Starting car 
[The issue with the car was solved: came with auto transmission from factory and the auto gearbox Neutral position confirmation wire was under the carpet and the wire was broken. The car couldn't start without the verification that the car was in neutral. We joined the wires back together and the no start was completly solved)

Here is a great bit of info and I can confirm that two of the colours are correct!!

Brown is ground - it is uncovered and stripped in the bundle going to the gauges

Grey and Red is a dimmer power 12 V - same as dash lights - will go up and down with wheel on dash. IT is stripped a bit in the bundle going to the switch and dimmer wheel.

If you are trying to decypher BMW wiring diagrams, it's rather essential to realise that:

BL=blue
BR=brown
GE=yellow
GN=green
GR=grey
OR=orange
RT=red
SW=black
VI=violet
WS=white

Some advice I received a long time ago from wiring guru Ron Styger of Unoffical BMW fame:

In general, a solid red wire is hot at all times.
A red wire with any color stripe is hot at all times after a fuse.
In general, a solid green wire is hot in run and start.
A green wire with any color stripe is hot in run and start after a fuse.
In general, a solid violet wire is hot in accessory, run and start.
A violet wire with any color stripe is hot in accessory, run and start after a fuse.
In general, brown is ground.
Gray with a red stripe is dimmed lighting.
I can't vouch for any of it myself, but Tim generally knows that of which he speaks. Hope it helps!